requirements.txt
MetaTrader5>=5.0.45
requests>=2.31.0
aiohttp>=3.9.5
python-socketio>=5.11.2
supabase>=2.4.5
pydantic>=2.7.1
python-dotenv>=1.0.1
